+++ app/src/main/java/com/app/testik/data/repository/FirestoreRepositoryImpl.kt
@@ -1,17 +1,22 @@
 package com.app.testik.data.repository
 
+import android.content.Context
 import com.app.testik.data.model.ApiResult
 import com.app.testik.data.model.RegistrationDto
 import com.app.testik.data.model.UserDto
 import com.app.testik.domain.repository.FirestoreRepository
 import com.app.testik.util.execute
+import com.app.testik.util.isOnline
+import com.app.testik.util.loadedFromServer
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.Source
+import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 
 class FirestoreRepositoryImpl @Inject constructor(
-    private val firebaseFirestore: FirebaseFirestore,
+    @ApplicationContext private val context: Context,
+    private val firebaseFirestore: FirebaseFirestore
 ): FirestoreRepository {
 
     override suspend fun addUser(data: RegistrationDto): ApiResult<Unit> {
@@ -52,13 +57,22 @@ class FirestoreRepositoryImpl @Inject constructor(
     }
 
     override suspend fun updateUser(data: UserDto): ApiResult<Unit> {
+        if (!isOnline(context)) return ApiResult.NoInternetError()
+
         return try {
             with(data) {
                 firebaseFirestore.collection("users").whereEqualTo("username", username).get().also {
                     it.await()
                     if (!it.result.isEmpty && it.result.documents.first().data?.get("email") != email) return ApiResult.Error("Username already taken")
                 }
-                firebaseFirestore.collection("users").document(email).set(data).execute()
+                val newData = mutableMapOf(
+                    "email" to email,
+                    "username" to username,
+                    "firstName" to firstName,
+                    "lastName" to lastName
+                )
+                if (avatar.loadedFromServer() || avatar.isEmpty()) newData["avatar"] = avatar
+                firebaseFirestore.collection("users").document(email).set(newData).execute()
             }
         } catch (e: Exception) {
             ApiResult.Error(e.message)
@@ -67,6 +81,7 @@ class FirestoreRepositoryImpl @Inject constructor(
 
     override suspend fun deleteUser(email: String?): ApiResult<Unit> {
         if (email == null) return ApiResult.Error("No email provided")
+        if (!isOnline(context)) return ApiResult.NoInternetError()
 
         return try {
             firebaseFirestore.collection("users").document(email).delete().execute()
