+++ app/src/main/java/com/app/testik/data/repository/StorageRepositoryImpl.kt
@@ -1,20 +1,26 @@
 package com.app.testik.data.repository
 
+import android.content.Context
 import android.net.Uri
 import androidx.core.net.toUri
 import com.app.testik.data.model.ApiResult
 import com.app.testik.data.model.UserDto
 import com.app.testik.domain.repository.StorageRepository
+import com.app.testik.util.isOnline
 import com.google.firebase.storage.FirebaseStorage
+import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.tasks.await
 import java.io.File
 import javax.inject.Inject
 
 class StorageRepositoryImpl @Inject constructor(
+    @ApplicationContext private val context: Context,
     private val firebaseStorage: FirebaseStorage
 ) : StorageRepository {
 
     override suspend fun uploadAvatar(data: UserDto): ApiResult<Uri> {
+        if (!isOnline(context)) return ApiResult.NoInternetError()
+
         return try {
             val uri = File(data.avatar).toUri()
             val avatar = firebaseStorage.reference.child("avatars").child(data.email)
@@ -27,6 +33,8 @@ class StorageRepositoryImpl @Inject constructor(
     }
 
     override suspend fun deleteAvatar(email: String): ApiResult<Unit> {
+        if (!isOnline(context)) return ApiResult.NoInternetError()
+
          try {
             firebaseStorage.reference.child("avatars").child(email).delete().also {
                 it.await()
